edit launch file

+++ dynamixel_workbench_single_manager/include/dynamixel_workbench_single_manager/dynamixel_workbench_single_manager.h
@@ -55,7 +55,7 @@ class DynamixelWorkbenchSingleManager
   // Parameters
   std::string device_name_;
   //uint64_t baud_rate_list_[BAUD_RATE_NUM];
-  float baud_rate_;
+  int baud_rate_;
   float protocol_version_;
   uint16_t dxl_model_number_;
   uint8_t dxl_model_id_;
